// src/app/pacient/models/PacientDto.ts
export interface PacientDto{

    pacientId: null;
    pacientName?: string;
    pacientDni: string;
    pacientAge: Number;
    pacientFono: string;
    pacientFono2: string;
    pacientAddress: string;
    pacientOccupation: string;
    pacientFecnac: string;
    pacientReferredBy: string;
    pacientEmail: string;
}